refactor(social): neutral provider registry breaks the IG↔TikTok cycle
- Move the canonical getSocialProvider switch into a neutral module
  (src/lib/sources/social-provider-registry.ts, next to social-provider.ts)
  that imports each platform's isConfigured probe + provider impl FROM its tree
  — one-directional, no ESM cycle. Previously instagram's registry owned the
  switch and imported {isTikTokConfigured, tiktokProvider} from tiktok's, which
  re-exported getSocialProvider back from instagram's (a deliberate cycle).
- Keep BOTH tree-local registries as thin re-export shims (now exporting
  isInstagramConfigured+instagramProvider / isTikTokConfigured+tiktokProvider
  and re-exporting getSocialProvider from the neutral module), so the ~9
  production import sites + the 21 per-tree vi.mock targets keep working
  unchanged. getSocialProvider sites stay on the shim path on purpose — that IS
  the mock seam (migrating them would bypass each test's vi.mock).
- Fix the imprecise 'live binding' comments: vi.mock replaces the module-graph
  node; the mocking tests use importOriginal()+spread (real probes survive) then
  override getSocialProvider — which is why tree isolation holds.
