+++ src/lib/sources/instagram/server/provider/registry.ts
@@ -1,26 +1,17 @@
-// Social-provider selection, keyed by platform via operator env (SOC-03).
+// Instagram-tree-local provider re-export (SOC-03).
 //
-// `getSocialProvider("instagram")` returns the configured provider impl, or
-// `null` when the operator hasn't wired one. The null path is the SOC-05
-// graceful-degrade hook: the not-configured walker + observability surfaces
-// read `null` as "feature off" rather than crashing. Empty INSTAGRAM_PROVIDER
-// (the safe default) ⇒ unresolved ⇒ degrade.
-//
-// This is the CANONICAL cross-platform switch — ONE switch, all platforms. Each
-// platform's config probe + impl live in that platform's own tree
-// (isTikTokConfigured + tiktokProvider come from the tiktok tree's registry
-// re-export); this file only routes by platform. New platforms (X) add one more
-// case keyed off the same `<PLATFORM>_PROVIDER` env convention.
+// Owns the Instagram config probe (isInstagramConfigured) + the provider impl
+// (instagramProvider) so the IG adapter folder reads its OWN accessor from
+// inside its own tree. The CANONICAL cross-platform `getSocialProvider` switch
+// lives in the NEUTRAL module (src/lib/sources/social-provider-registry.ts —
+// one switch, all platforms; do NOT fork it); that module imports the two
+// bindings above FROM here (neutral → tree, one-directional, no cycle). This
+// file re-exports getSocialProvider from there as a thin shim so the IG tree's
+// import sites + the per-tree vi.mock targets keep working unchanged.
 
 import { env } from "$lib/server/config/env.js";
 import { scrapeCreatorsProvider } from "./scrapecreators.js";
-import {
-  isTikTokConfigured,
-  tiktokProvider,
-} from "$lib/sources/tiktok/server/provider/registry.js";
-import type { SocialPlatform, SocialProvider } from "$lib/sources/social-provider.js";
-
-export { isTikTokConfigured };
+import type { SocialProvider } from "$lib/sources/social-provider.js";
 
 /** Whether the operator has a usable Instagram provider configured: the
  *  provider is selected AND its credential is non-empty. */
@@ -28,21 +19,16 @@ export function isInstagramConfigured(): boolean {
   return env.INSTAGRAM_PROVIDER === "scrapecreators" && env.SCRAPECREATORS_API_KEY !== "";
 }
 
-/**
- * Resolve the provider for a platform, or `null` when not configured.
- *
- * `null` is a first-class outcome (SOC-05) — the caller degrades gracefully
- * (the walker no-ops, the UI shows "not configured") rather than treating it
- * as an error.
- */
-export function getSocialProvider(platform: SocialPlatform): SocialProvider | null {
-  switch (platform) {
-    case "instagram":
-      return isInstagramConfigured() ? scrapeCreatorsProvider : null;
-    case "tiktok":
-      return isTikTokConfigured() ? tiktokProvider : null;
-    default:
-      // Twitter / X provider lands in Phase 11.
-      return null;
-  }
-}
+/** The Instagram provider impl (the SOC-02 issuer). */
+export const instagramProvider: SocialProvider = scrapeCreatorsProvider;
+
+// Re-export the CANONICAL cross-platform switch from the neutral registry (the
+// single switch, all platforms — do NOT fork it). The IG tree's handlers /
+// adapter import `getSocialProvider` from HERE so they read their provider
+// accessor from inside their own folder; the routing logic lives in one
+// neutral place. Integration tests vi.mock THIS module to inject a fake
+// provider at the seam: vi.mock replaces this module's node in the graph, and
+// the mocks spread importOriginal() (so the real isInstagramConfigured /
+// instagramProvider survive) then override getSocialProvider — which is why a
+// mock of the IG registry never leaks into the TikTok tree's own copy.
+export { getSocialProvider } from "$lib/sources/social-provider-registry.js";
