+++ src/lib/sources/tiktok/server/provider/registry.ts
@@ -1,11 +1,13 @@
 // TikTok-tree-local provider re-export (SOC-03).
 //
-// Thin re-export of `isTikTokConfigured` + the provider so the tiktok adapter
-// folder imports its OWN config probe + impl from inside its own tree, instead
-// of reaching across into instagram's registry for them. The CANONICAL
-// cross-platform `getSocialProvider` switch lives in instagram's registry today
-// (one switch, all platforms — do NOT fork it); this file only surfaces the
-// tiktok-specific pieces for the tiktok tree's local imports.
+// Owns the TikTok config probe (isTikTokConfigured) + the provider impl
+// (tiktokProvider) so the TikTok adapter folder reads its OWN accessor from
+// inside its own tree. The CANONICAL cross-platform `getSocialProvider` switch
+// lives in the NEUTRAL module (src/lib/sources/social-provider-registry.ts —
+// one switch, all platforms; do NOT fork it); that module imports the two
+// bindings above FROM here (neutral → tree, one-directional, no cycle). This
+// file re-exports getSocialProvider from there as a thin shim so the TikTok
+// tree's import sites + the per-tree vi.mock targets keep working unchanged.
 
 import { env } from "$lib/server/config/env.js";
 import { scrapeCreatorsTikTokProvider } from "./scrapecreators-tiktok.js";
@@ -21,12 +23,13 @@ export function isTikTokConfigured(): boolean {
 /** The TikTok provider impl (the SOC-02 issuer). */
 export const tiktokProvider: SocialProvider = scrapeCreatorsTikTokProvider;
 
-// Re-export the CANONICAL cross-platform switch from instagram's registry (the
-// single switch, all platforms — do NOT fork it). The tiktok tree's handlers /
+// Re-export the CANONICAL cross-platform switch from the neutral registry (the
+// single switch, all platforms — do NOT fork it). The TikTok tree's handlers /
 // adapter import `getSocialProvider` from HERE so they read their provider
-// accessor from inside their own folder; the routing logic still lives in one
-// place. Integration tests mock THIS module's getSocialProvider to inject a fake
-// tiktok provider at the seam (mirrors the instagram tests). The re-export is a
-// live binding, so a vi.mock of this module shadows it for the tiktok tree
-// without touching the instagram tree's own copy.
-export { getSocialProvider } from "$lib/sources/instagram/server/provider/registry.js";
+// accessor from inside their own folder; the routing logic lives in one
+// neutral place. Integration tests vi.mock THIS module to inject a fake tiktok
+// provider at the seam: vi.mock replaces this module's node in the graph, and
+// the mocks spread importOriginal() (so the real isTikTokConfigured /
+// tiktokProvider survive) then override getSocialProvider — which is why a mock
+// of the TikTok registry never leaks into the instagram tree's own copy.
+export { getSocialProvider } from "$lib/sources/social-provider-registry.js";
